Update EnrichUrlClient to use new property and add TLS support
The code has been updated to use the new property 'folio.client.okapiUrl' in EnrichUrlClient and replace the usage of the deprecated 'okapi_url'. Moreover, an enhancement has been made to support TLS; this includes modifications to the FeignClientConfiguration and pom.xml files, as well as creation of the EdgeFeignClientProperties class.